Revert "create base for go rewrite"
This reverts commit adc05bdb56e811efa9707629ca7362400fb11ad2.

+++ migrations/10.sql
@@ -1,6 +1,6 @@
 CREATE TABLE tournament_staff(
 	id INT NOT NULL,
-	tournament BIGINT NOT NULL,
+	tournament INT NOT NULL,
 	privileges INT NOT NULL,
 	PRIMARY KEY(id, tournament)
 );
